+++ src/lsb_private.h
@@ -32,7 +32,7 @@ struct lua_sandbox {
   lsb_state       state;
   lsb_output_data output;
   char*           lua_file;
-  unsigned        usage[LSB_UT_MAX][LSB_US_MAX];
+  size_t          usage[LSB_UT_MAX][LSB_US_MAX];
   char            error_message[LSB_ERROR_SIZE];
 };
